fix(PX-1342): pr feedbacks

+++ packages/jira/src/util/delete-process.ts
@@ -4,7 +4,11 @@ import { RetryTableMapping } from 'src/model/retry-table-mapping';
 
 const dynamodbClient = DynamoDbDocClient.getInstance();
 
-export async function deleteProcessfromDdb(processId: string): Promise<void> {
+export async function deleteProcessfromDdb(processId: string | undefined): Promise<void> {
+  if (!processId) {
+    return;
+  }
+  logger.info('deleting_process_from_DDB', { processId });
   await dynamodbClient.delete(new RetryTableMapping().prepareDeleteParams(processId));
   logger.info('RetryProcessHandlerProcess.delete', { processId });
 }
